+++ KeePassiumLib/KeePassiumLib/db/kp2/Database2.swift
@@ -703,6 +703,7 @@ public class Database2: Database {
                     comment: "A warning about unused attachments after loading the database. [lastUsedAppName: String]"),
                 lastUsedAppName)
             warnings.messages.append(warningMessage)
+            warnings.isGeneratorImportant = true
             
             let unusedIDs = unusedBinaries
                 .map { String($0) }
@@ -780,6 +781,7 @@ public class Database2: Database {
             listOfEntryNames)
         Diag.warning(warningMessage)
         warnings.messages.append(warningMessage)
+        warnings.isGeneratorImportant = true
     }
     
     private func checkCustomFieldsIntegrity(allEntries: [Entry], warnings: DatabaseLoadingWarnings) {
@@ -802,6 +804,7 @@ public class Database2: Database {
                 comment: "A warning about misformatted custom fields after loading the database. [entryPaths: String]"),
             entryPaths)
         warnings.messages.append(warningMessage)
+        warnings.isGeneratorImportant = true
     }
         
     private func updateBinaries(root: Group2) {