+++ KeePassiumLib/KeePassiumLib/util/ProcessInfo+extensions.swift
@@ -1,0 +1,17 @@
+//  KeePassium Password Manager
+//  Copyright © 2021 Andrei Popleteev <info@keepassium.com>
+//
+//  This program is free software: you can redistribute it and/or modify it
+//  under the terms of the GNU General Public License version 3 as published
+//  by the Free Software Foundation: https://www.gnu.org/licenses/).
+//  For commercial licensing, please contact the author.
+
+extension ProcessInfo {
+    
+    public static var isRunningOnMac: Bool {
+        guard #available(iOS 14, *) else {
+            return false
+        }
+        return ProcessInfo.processInfo.isiOSAppOnMac
+    }
+}